Changed array subscripts in do_pose and boot_pose_messages to integers.

+++ act.social.c
@@ -283,9 +283,9 @@ void do_insult (struct char_data *ch, char *argument, int cmd)
 void boot_pose_messages (void)
 {
   FILE *fl;
-  byte counter;
+  int counter;
   int tmp;
-  byte class;
+  int class;
 
   if (!(fl = fopen (POSEMESS_FILE, "r"))) {
     perror ("boot_pose_messages");
@@ -308,8 +308,8 @@ void boot_pose_messages (void)
 
 void do_pose (struct char_data *ch, char *argument, int cmd)
 {
-  byte to_pose;
-  byte counter;
+  int to_pose;
+  int counter;
 
   if ((GET_LEVEL (ch) < pose_messages[0].level) || IS_NPC (ch)) {
     send_to_char ("You can't do that.\n\r", ch);
